Explicitly cast $lineNumber to int in GcovTar handler
This fixes the following error that occurs while parsing some
GcovTar submissions:

TypeError: Unsupported operand types: string - int
/cdash/app/cdash/xml_handlers/GcovTar_handler.php:322

+++ app/cdash/xml_handlers/GcovTar_handler.php
@@ -292,7 +292,7 @@ class GCovTarHandler extends NonSaxHandler
             if (count($fields) > 2) {
                 // Separate out delimited values from this line.
                 $timesHit = trim($fields[0]);
-                $lineNumber = trim($fields[1]);
+                $lineNumber = (int) trim($fields[1]);
                 $sourceLine = rtrim($fields[2]);
 
                 //check for duplicate line output
